+++ scripts/actions/session.js
@@ -4,6 +4,7 @@ import {
   SESSION_SETUP_COMPLETE,
   SESSION_SERVERINFO_SUCCESS,
   SESSION_SERVERINFO_FAILURE,
+  SESSION_BUCKETS_REQUEST,
   SESSION_BUCKETS_SUCCESS,
   SESSION_BUCKETS_FAILURE,
   SESSION_LOGOUT,
@@ -28,6 +29,10 @@ export function serverInfoSuccess(serverInfo) {
 
 export function serverInfoFailure(error) {
   return {type: SESSION_SERVERINFO_FAILURE, error};
+}
+
+export function listBuckets() {
+  return {type: SESSION_BUCKETS_REQUEST};
 }
 
 export function bucketsSuccess(buckets) {
