+++ scripts/constants.js
@@ -1,5 +1,3 @@
-export const CLIENT_BUSY = "CLIENT_BUSY";
-
 export const COLLECTION_RESET = "COLLECTION_RESET";
 export const COLLECTION_LOAD_REQUEST = "COLLECTION_LOAD_REQUEST";
 export const COLLECTION_LOAD_SUCCESS = "COLLECTION_LOAD_SUCCESS";
@@ -9,7 +7,6 @@ export const COLLECTION_UPDATE_SUCCESS = "COLLECTION_UPDATE_SUCCESS";
 export const COLLECTION_DELETE_REQUEST = "COLLECTION_DELETE_REQUEST";
 export const COLLECTION_RECORDS_REQUEST = "COLLECTION_RECORDS_REQUEST";
 export const COLLECTION_RECORDS_SUCCESS = "COLLECTION_RECORDS_SUCCESS";
-export const COLLECTION_RECORD_CREATED = "COLLECTION_RECORD_CREATED";
 
 export const NOTIFICATION_ADDED = "NOTIFICATION_ADDED";
 export const NOTIFICATION_REMOVED = "NOTIFICATION_REMOVED";
