+++ scripts/containers/AddFormPage.js
@@ -3,7 +3,6 @@ import { connect } from "react-redux";
 import { updatePath } from "redux-simple-router";
 
 import AddForm from "../components/AddForm";
-import * as ClientActions from "../actions/client";
 import * as CollectionActions from "../actions/collection";
 import * as NotificationsActions from "../actions/notifications";
 
@@ -18,7 +17,6 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
   return bindActionCreators({
     ...CollectionActions,
-    ...ClientActions,
     ...NotificationsActions,
     updatePath
   }, dispatch);
