+++ scripts/containers/CollectionListPage.js
@@ -1,7 +1,6 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import CollectionList from "../components/CollectionList";
-import * as ClientActions from "../actions/client";
 import * as CollectionActions from "../actions/collection";
 import * as NotificationsActions from "../actions/notifications";
 import { updatePath } from "redux-simple-router";
@@ -15,7 +14,6 @@ function mapStateToProps(state) {
 
 function mapDispatchToProps(dispatch) {
   return bindActionCreators({
-    ...ClientActions,
     ...CollectionActions,
     ...NotificationsActions,
     updatePath
